logfs: use uintptr_t to hold opaque pointers
When storing an opaque pointer value in a uint, the
appropriate type is uintptr_t which is guaranteed to
be wide enough to hold a pointer.

This is particularly important when the code can be
compiled in a sim or unit test environment which is
intended to run on a 64-bit machine.

+++ flight/targets/DiscoveryF4/System/pios_board.c
@@ -112,7 +112,7 @@ void PIOS_Board_Init(void) {
 	/* Connect flash to the approrpiate interface and configure it */
 	uint32_t flash_id;
 	PIOS_Flash_Internal_Init(&flash_id, &flash_internal_cfg);
-	uint32_t fs_id;
+	uintptr_t fs_id;
 	PIOS_FLASHFS_Logfs_Init(&fs_id, &flashfs_internal_cfg, &pios_internal_flash_driver, flash_id);
 #endif
